requirements.txt
flask
transformers
torch
sentence-transformers
faiss-cpu

